requirements.txt
chromadb==0.5.23
fastapi==0.124.4
httpx<0.28
jinja2>=3.1.6
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
openai==1.52.2
overrides>=7.7.0
pydantic==2.10.6
pypdf==5.4.0
python-docx==1.1.2
python-multipart==0.0.17
sentence-transformers==3.2.1
uvicorn==0.33.0
